% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fx.R
\name{effort_F_plot}
\alias{effort_F_plot}
\title{effort_F_plot}
\usage{
effort_F_plot(
  E_t,
  F_ta,
  vul_age,
  x_min,
  x_max,
  y_min,
  y_max,
  years,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL
)
}
\description{
effort_F_plot
}
