% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_data}
\alias{read_data}
\title{read_data()
function to read in data files in .dat format
.dat file
first line are details
second line are years
third lines are ages
fourth line is data type
\itemize{
\item 1: matrix
\item 2: vector
}}
\usage{
read_data(file)
}
\description{
read_data()
function to read in data files in .dat format
.dat file
first line are details
second line are years
third lines are ages
fourth line is data type
\itemize{
\item 1: matrix
\item 2: vector
}
}
