% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_table_nowarn}
\alias{read_table_nowarn}
\title{read_table_nowarn
function to supress incomplete final lines warning}
\usage{
read_table_nowarn(...)
}
\description{
read_table_nowarn
function to supress incomplete final lines warning
}
