% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{rtmb_template}
\alias{rtmb_template}
\title{rtmb_template}
\usage{
rtmb_template(
  name,
  wd = NULL,
  random = NULL,
  control = NULL,
  newton_steps = FALSE,
  jitter_test = FALSE,
  likeprof = FALSE,
  mcmc = FALSE
)
}
\description{
rtmb_template
}
