% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{setup_data}
\alias{setup_data}
\title{setup_data()
function to set up data for model}
\usage{
setup_data(
  data_list,
  type,
  fishery,
  maturity_at_age = NULL,
  weight_at_age = NULL,
  natural_mortality = NULL
)
}
\description{
setup_data()
function to set up data for model
}
