% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fx.R
\name{uncertainty_plot}
\alias{uncertainty_plot}
\title{uncertainty_plot}
\usage{
uncertainty_plot(
  vec,
  vec_sd,
  years,
  quantile = 0.975,
  y_lab = NULL,
  title = NULL
)
}
\description{
uncertainty_plot
}
