% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fx.R
\name{year_age_b_plot}
\alias{year_age_b_plot}
\title{year_age_b_plot}
\usage{
year_age_b_plot(
  df,
  years,
  ages,
  title = NULL,
  y_lab = NULL,
  legend = FALSE,
  legend_title = NULL
)
}
\description{
year_age_b_plot
}
